use fabfile.py, maybe too old but better known for the moment. add fabfile.pyc to .gitignore

+++ experiments/fabfile.py
@@ -7,13 +7,9 @@ Author: Nathan Flick
 import os
 import os.path
 import datetime
-import re
-from time import sleep
 
 import fabric.api as fab
 import fabric.utils
-
-fab.env.user = 'root'
 
 DEFAULT_MAC = 'tdma-4.txt'
 
@@ -52,54 +48,35 @@ def load_b43():
         pass
 
 @fab.task
-@fab.runs_once
-def setup(branch='metamac', firmware_branch=None, debug=False):
-    fab.local('rm -rf cache && mkdir cache')
-    with fab.lcd('cache'):
-        fab.local('wget http://launchpadlibrarian.net/167352137/libxml2_2.9.1+dfsg1-3ubuntu4_i386.deb')
-        fab.local('wget http://launchpadlibrarian.net/167352140/libxml2-dev_2.9.1+dfsg1-3ubuntu4_i386.deb')
-        fab.local('wget github.com/nflick/wireless-mac-processor/archive/{0}.zip'.format(branch))
-        fab.local('wget -O iperf.zip github.com/nflick/iperf/archive/master.zip')
-        if firmware_branch is not None and firmware_branch != branch:
-            fab.local('wget github.com/nflick/wireless-mac-processor/archive/{0}.zip'.format(firmware_branch))
-    fab.execute(setup_remote, branch, firmware_branch, debug)
-
-@fab.task
 @fab.parallel
-def setup_remote(branch='metamac', firmware_branch=None, debug=False):
+def setup(branch='metamac', get_src=True,firmware_branch=None, debug=False,gituser='fabriziogiuliano'):
+
     '''Sets up the node by downloading the specified branch or commit and extracting necessary
     files, installing the WMP firmware, and building needed tools.
     '''
 
     if firmware_branch is None:
         firmware_branch = branch
+    if get_src:
+	    with fab.cd('/tmp'):
+		fab.run('rm -f *.deb')
+		fab.run('wget http://security.ubuntu.com/ubuntu/pool/main/libx/libxml2/libxml2_2.9.1+dfsg1-3ubuntu4.7_i386.deb')
+		fab.run('wget http://security.ubuntu.com/ubuntu/pool/main/libx/libxml2/libxml2-dev_2.9.1+dfsg1-3ubuntu4.7_i386.deb')
+		fab.run('dpkg -i *.deb')
+		fab.run('rm -f *.deb')
 
-    fab.run('rm -f *.deb')
-    fab.put(local_path='cache/libxml2_2.9.1+dfsg1-3ubuntu4_i386.deb', remote_path='.')
-    fab.put(local_path='cache/libxml2-dev_2.9.1+dfsg1-3ubuntu4_i386.deb', remote_path='.')
-    fab.run('dpkg -i *.deb')
-    fab.run('rm -f *.deb')
+	    fab.run('rm -rf metamac && mkdir metamac')
+	    with fab.cd('metamac'):
+		fab.run('wget github.com/{1}/wireless-mac-processor/archive/{0}.zip'.format(branch,gituser))
+		fab.run('unzip {0}.zip "wireless-mac-processor-{0}/wmp-injection/bytecode-manager/*"'.format(branch))
+		fab.run('unzip {0}.zip "wireless-mac-processor-{0}/mac-programs/metaMAC-program/*"'.format(branch))
+	
+		if firmware_branch != branch:
+		    fab.run('rm {0}.zip'.format(branch))
+		    fab.run('wget github.com/{1}/wireless-mac-processor/archive/{0}.zip'.format(firmware_branch,gituser))
 
-    fab.run('rm -rf iperf-master')
-    fab.put(local_path='cache/iperf.zip', remote_path='.')
-    fab.run('unzip iperf.zip')
-    fab.run('rm iperf.zip')
-    with fab.cd('iperf-master'):
-        fab.run('./configure')
-        fab.run('make')
-
-    fab.run('rm -rf metamac && mkdir metamac')
-    with fab.cd('metamac'):
-        fab.put(local_path='cache/{0}.zip'.format(branch), remote_path='.')
-        fab.run('unzip {0}.zip "wireless-mac-processor-{0}/wmp-injection/bytecode-manager/*"'.format(branch))
-        fab.run('unzip {0}.zip "wireless-mac-processor-{0}/mac-programs/metaMAC-program/*"'.format(branch))
-
-        if firmware_branch != branch:
-            fab.run('rm {0}.zip'.format(branch))
-            fab.put(local_path='cache/{0}.zip'.format(firmware_branch), remote_path='.')
-
-        fab.run('unzip {0}.zip "wireless-mac-processor-{0}/wmp-engine/broadcom-metaMAC/*"'.format(firmware_branch))
-        fab.run('rm {0}.zip'.format(firmware_branch))
+		fab.run('unzip {0}.zip "wireless-mac-processor-{0}/wmp-engine/broadcom-metaMAC/*"'.format(firmware_branch))
+		fab.run('rm {0}.zip'.format(firmware_branch))
 
     with fab.cd('metamac/wireless-mac-processor-{0}/wmp-engine/broadcom-metaMAC/'.format(firmware_branch)):
         fab.run('cp ucode5.fw b0g0bsinitvals5.fw b0g0initvals5.fw /lib/firmware/b43')
@@ -108,17 +85,61 @@ def setup_remote(branch='metamac', firmware_branch=None, debug=False):
         fab.run('sleep 1')
         fab.run('modprobe b43 qos=0')
         fab.run('sleep 1')
-
     with fab.cd('~/metamac/wireless-mac-processor-{0}/wmp-injection/bytecode-manager/'.format(branch)):
         if debug:
-            fab.run("sed -i 's/CFLAGS=/CFLAGS=-g /' Makefile") # Add -g flag to Makefile
-            fab.run("sed -i 's/-O[0-9]/ /' Makefile")          # Remove optimization flag from Makefile
+            fab.run("sed -i 's/CFLAGS=/CFLAGS=-g /' Makefile")
+            fab.run("sed -i 's/-O[0-9]/ /' Makefile")
         fab.run('make bytecode-manager')
         fab.run('make metamac')
         fab.run('make tsfrecorder')
-        fab.run('cp bytecode-manager ~/metamac/')
-        fab.run('cp metamac ~/metamac/')
-        fab.run('cp tsfrecorder ~/metamac/')
+        fab.run('cp -f bytecode-manager ~/metamac/')
+        fab.run('cp -f metamac ~/metamac/')
+        fab.run('cp -f tsfrecorder ~/metamac/')
+@fab.task
+@fab.parallel
+def setup_local(branch='metamac', get_src=True,firmware_branch=None, debug=False):
+
+    '''Sets up the node by downloading the specified branch or commit and extracting necessary
+    files, installing the WMP firmware, and building needed tools.
+    '''
+
+    if firmware_branch is None:
+        firmware_branch = branch
+    if get_src:
+	    with fab.cd('/tmp'):
+		fab.run('rm -f *.deb')
+		fab.run('wget http://security.ubuntu.com/ubuntu/pool/main/libx/libxml2/libxml2_2.9.1+dfsg1-3ubuntu4.7_i386.deb')
+		fab.run('wget http://security.ubuntu.com/ubuntu/pool/main/libx/libxml2/libxml2-dev_2.9.1+dfsg1-3ubuntu4.7_i386.deb')
+		fab.run('dpkg -i *.deb')
+		fab.run('rm -f *.deb')
+
+	    fab.run('rm -rf metamac && mkdir metamac')
+	    with fab.cd('metamac'):
+		wmp_dir='wireless-mac-processor-{0}/'.format(branch);
+	    	fab.run('mkdir {0} && mkdir {0}/wmp-injection'.format(wmp_dir));
+		fab.put(local_path='git/wireless-mac-processor/wmp-injection/*',remote_path='wireless-mac-processor-{0}/wmp-injection/'.format(branch));
+		
+	    	fab.run('mkdir {0}/mac-programs && mkdir {0}/mac-prorams/metaMAC-program/'.format(wmp_dir));
+		fab.put(local_path='git/wireless-mac-processor/mac-programs/metaMAC-program/*',remote_path='wireless-mac-processor-{0}/mac-programs/metaMAC-program/'.format(branch) );
+
+
+    fab.put(local_path='git/meta-MAC/wmp-firmware/wmp3-2.28/*.fw',remote_path='/lib/firmware/b43/');
+    with fab.settings(warn_only=True):
+    	fab.run('rmmod b43')
+    fab.run('sleep 1')
+    fab.run('modprobe b43 qos=0')
+    fab.run('sleep 1')
+
+    with fab.cd('~/metamac/wireless-mac-processor-{0}/wmp-injection/bytecode-manager/'.format(branch)):
+        if debug:
+            fab.run("sed -i 's/CFLAGS=/CFLAGS=-g /' Makefile")
+            fab.run("sed -i 's/-O[0-9]/ /' Makefile")
+        fab.run('make bytecode-manager')
+        fab.run('make metamac')
+        fab.run('make tsfrecorder')
+        fab.run('cp -f bytecode-manager ~/metamac/')
+        fab.run('cp -f metamac ~/metamac/')
+        fab.run('cp -f tsfrecorder ~/metamac/')
 
 @fab.task
 @fab.parallel
@@ -132,12 +153,6 @@ def ap_ify(path):
 def on_node(node):
     return fab.env.host_string.split('@')[-1] == node
 
-def node_num(node):
-    m = re.match(r'alix(\d+)', node)
-    if not m:
-        raise Exception('Unknown node naming convention')
-    return int(m.group(1))
-
 @fab.task
 def load_mac(mac, ap_node=None):
     '''Loads the given MAC protocol onto the WMP firmware. Paths are relative to the
@@ -150,93 +165,107 @@ def load_mac(mac, ap_node=None):
     fab.run('metamac/bytecode-manager -a 1')
 
 @fab.task
-def start_ap(mac, channel=6):
-    fab.run('ifconfig wlan0 192.168.0.$(hostname | grep -Eo [0-9]+) netmask 255.255.255.0 up')
-    load_mac(mac)
+def start_ap(mac):
+    #fab.run('ifconfig wlan0 192.168.0.$(hostname | grep -Eo [0-9]+) netmask 255.255.255.0 up')
+    #load_mac(mac)
     with fab.settings(warn_only=True):
-        fab.run('killall hostapd')
-    fab.run('sleep 3')
+        fab.run('killall -9 hostapd')
+    	fab.run('killall -9 metamac')
+    fab.run('scp ~/work/openfwwf-5.2/*.fw /lib/firmware/b43/')
     with fab.cd('work/association'):
         fab.run('if ! grep basic_rates=60 hostapd.conf; then echo "basic_rates=60" >> hostapd.conf; fi')
         fab.run('if ! grep supported_rates=60 hostapd.conf; then echo "supported_rates=60" >> hostapd.conf; fi')
         fab.run("sed -i 's/macaddr_acl=1/macaddr_acl=0/' hostapd.conf")
-        fab.run("sed -i 's/channel=6/channel={0}/' hostapd.conf".format(channel))
+        fab.run("sed -i 's/channel=11/channel=6/' hostapd.conf")
     # Must not be prefixed with cd work/association or else the cd will interfere with nohup.
     fab.run('nohup work/association/up-hostapd.sh work/association/hostapd.conf 192.168.0.$(hostname | grep -Eo [0-9]+) >& hostapd.log < /dev/null &', pty=False)
-    fab.run('sleep 8')
-    load_mac(mac)
+    fab.run('sleep 4')
+    #load_mac(mac)
 
 @fab.task
-def associate(mac):
-    with fab.settings(warn_only=True):
-        fab.run('killall hostapd')
-    fab.run('sleep 3')
-    fab.run('ifconfig wlan0 192.168.0.$(hostname | grep -Eo [0-9]+) netmask 255.255.255.0 up')
-    load_mac(mac)
-    with fab.cd('work/association'):
-        attempts = 0
-        result = fab.run('./ass.sh alix-ap 192.168.0.$(hostname | grep -Eo [0-9]+)')
-        while 'iwconfig result Access' not in result and attempts < 4:
-            result = fab.run('./ass.sh alix-ap 192.168.0.$(hostname | grep -Eo [0-9]+)')
-            attempts += 1
-        if attempts >= 4:
-            fabric.utils.abort('Could not associate node {0} with access point.'.format(fab.env.host_string))
-    load_mac(mac)
+@fab.parallel
+def associate(mac=DEFAULT_MAC):
+
+    #load_mac(mac)
+    fab.run('modprobe b43 qos=0')
+    fab.run('rmmod b43')
+    fab.run('modprobe b43 qos=0')
+    fab.run('ifconfig wlan0 192.168.0.$(hostname | grep -Eo [0-9]+) netmask 255.255.255.0')
+    fab.run('iwconfig wlan0 essid alix-ap')
+    fab.run('iwconfig wlan0 txpower 20dbm')
+    result = fab.run('iwconfig wlan0 | grep Access | awk \'{ print $4 }\';')
+    attempts = 0
+    while 'Not-Associated' in result and attempts < 10:
+        fab.run('iwconfig wlan0 essid alix-ap')
+        fab.run('sleep 1')
+        result = fab.run('iwconfig wlan0 | grep Access | awk \'{ print $4 }\';')
+        attempts += 1
+    if attempts >= 10:
+        fabric.utils.abort('Could not associate node {0} with access point.'.format(fab.env.host_string))
+    #load_mac(mac)
 
 @fab.task
 @fab.runs_once
 def network(ap_node, mac=DEFAULT_MAC):
+
+    fab.execute(kill_metamac)
     fab.execute(start_ap, ap_ify(mac), hosts=[h for h in fab.env.hosts if h.split('@')[-1] == ap_node])
     fab.execute(associate, mac, hosts=[h for h in fab.env.hosts if h.split('@')[-1] != ap_node])
 
 @fab.task
-@fab.parallel
-def start_iperf_server(ports):
+def start_iperf_server():
     with fab.settings(warn_only=True):
-        fab.run('killall iperf-master/src/iperf3')
+        fab.run('killall iperf')
     fab.local('sleep 3')
-    for port in ports:
-        fab.run('nohup iperf-master/src/iperf3 -s -p {0} > iperf.out 2> iperf.err < /dev/null &'.format(port), pty=False)
+    fab.run('nohup iperf -i 1 -s -u -y C > iperf.out 2> iperf.err < /dev/null &', pty=False)
 
 @fab.task
-@fab.parallel
-def run_iperf_client(ap_node, duration, linear=False, bandwidth=6000000, delay=False):
-    if on_node(ap_node):
-        target = next(h.split('@')[-1] for h in fab.env.hosts if h.split('@')[-1] != ap_node)
-    else:
-        target = ap_node
-    port = 5000 + node_num(fab.env.host_string.split('@')[-1])
-    for i, host in enumerate(fab.env.hosts):
-        if host.split('@')[-1] == fab.env.host_string.split('@'):
-            delay = i * 0.5
-    else:
-        delay = 0.0
-    sleep(delay)
-    fab.run('iperf-master/src/iperf3 -c 192.168.0.$(echo {0} | grep -Eo [0-9]+) -p {1} -u -t {2} -b {3} {4}'.format(
-        target, port, duration, bandwidth, '-r' if linear else ''))
-
-@fab.task
-@fab.parallel
-def start_metamac(suite, ap_node=None, cycle=False, eta=0.0):
+def stop_iperf_server():
     with fab.settings(warn_only=True):
-        fab.run('killall -s INT metamac/metamac')
-    fab.run('sleep 2')
+        fab.run('killall iperf')
+
+@fab.task
+@fab.parallel
+def stop_iperf():
+    with fab.settings(warn_only=True):
+        fab.run('killall -9 iperf')
+
+@fab.task
+@fab.parallel
+def stop_hostapd():
+    with fab.settings(warn_only=True):
+        fab.run('killall -9 hostapd')
+
+@fab.task
+@fab.parallel
+def stop_wifi():
+    with fab.settings(warn_only=True):
+        fab.run('rmmod b43')
+
+@fab.task
+@fab.parallel
+def run_iperf_client(server, duration):
+    fab.run('iperf -c 192.168.0.$(echo {0} | grep -Eo [0-9]+) -u -t {1} -b 6000000'.format(server, duration))
+
+@fab.task
+def start_metamac(suite, ap_node=None, cycle=False):
     if on_node(ap_node):
         suite = ap_ify(suite)
-    arguments = ''
-    if cycle:
-        arguments += '-c '
-    if eta > 0.0:
-        arguments += '-e {0} '.format(eta)
-    fab.run('nohup metamac/metamac {0} -l metamac.log metamac/wireless-mac-processor-*/mac-programs/metaMAC-program/{1} > metamac.out 2> metamac.err < /dev/null &'.format(arguments, suite), pty=False)
+    if not on_node(ap_node):
+    	fab.run('killall -9 metamac; nohup metamac/metamac {0} -l metamac.log metamac/wireless-mac-processor-*/mac-programs/metaMAC-program/{1} > metamac.out 2> metamac.err < /dev/null &'.format('-c' if cycle else '', suite), pty=False)
     fab.run('sleep 2')
 
+@fab.task
+@fab.parallel
+def kill_metamac():
+    with fab.settings(warn_only=True):
+    	fab.run('killall -9 metamac')
 @fab.task
 @fab.parallel
 def stop_metamac(trialnum):
     fab.run('sleep 2')
     with fab.settings(warn_only=True):
-        fab.run('killall -s INT metamac/metamac')
+    	fab.run('killall -9 metamac')
     fab.local('sleep 2')
     localname = 'data/{0}-{1}-{2}.csv'.format(datetime.date.today(), fab.env.host_string.split('@')[-1], trialnum)
     fab.local('mkdir -p data')
@@ -245,13 +274,51 @@ def stop_metamac(trialnum):
 
 @fab.task
 @fab.runs_once
-def run_trial(trialnum, suite, ap_node, linear=False, bandwidth=6000000, delay=False):
-    ports = [str(5000 + node_num(h.split('@')[-1])) for h in fab.env.hosts]
-    fab.execute(start_iperf_server, ports, hosts=[ap_node])
+def pkt_dump(trialnum,action='run'):
+	fab.local('sudo tcpdump -i wlan0 > /dev/null &')
+	fab.local('sleep 1')
+	fab.local('sudo killall -9 tcpdump')
+	fab.local('sudo rmmod ath9k')
+	if action=='run':
+		fab.local('sudo modprobe ath9k')
+		fab.local('sudo rfkill unblock all')
+		fab.local('sudo ifconfig wlan0 down; sudo iwconfig wlan0 mode monitor; sudo ifconfig wlan0 up; sudo iwconfig wlan0 channel 6; sudo iwconfig wlan0 rate 6M fixed')
+		localname = 'data/{0}-pkt_dump-{1}.csv'.format(datetime.date.today(), trialnum)
+		fab.local('mkdir -p data')
+		fab.local('sleep 5')
+		#fab.local('sudo tcpdump -i wlan0 net 192.168.0.0/24 and dst port 5001 | awk \'{print $2","$14}\' | sed \'s/us//\' > '+localname+'&')
+		fab.local('sudo tcpdump -i wlan0 | grep -v Ack | grep -v Beacon | grep \"IP 192.168.0\" | awk \'{print $1","$2","$14}\' | sed \'s/us//\' > '+localname+' &')
+
+
+
+@fab.task
+@fab.runs_once
+def run_trial(trialnum, suite, ap_node):
+    fab.execute(start_iperf_server, hosts=[ap_node])
+    fab.execute(kill_metamac)
     fab.execute(start_metamac, suite, ap_node)
-    fab.execute(run_iperf_client, ap_node, 30, linear=linear, bandwidth=bandwidth, delay=delay)
+    fab.execute(pkt_dump,trialnum)
+    fab.execute(run_iperf_client, ap_node, 60, hosts=[h for h in fab.env.hosts if h.split('@')[-1] != ap_node])
     fab.execute(stop_metamac, trialnum)
+    fab.execute(stop_iperf_server, hosts=[ap_node])
+
+
+
+@fab.task
+def sync_date():
+    controller_time=fab.local('echo $(date +%T)',capture=True)
+    #fab.local('echo {0}'.format(controller_time))
+    fab.run('date +%T -s {0}'.format(controller_time))
+
+@fab.task
+@fab.runs_once
+def stop_all():
+    fab.execute(stop_wifi)
+    fab.execute(kill_metamac)
+    fab.execute(stop_iperf)
+    fab.execute(stop_hostapd)
 
 @fab.task
 def trial_errors():
     fab.run('cat metamac.err')
+
